+++ include/power.h
@@ -1,4 +1,0 @@
-#pragma once
-
-// Trigger an immedate reboot
-void power_reboot();