+++ include/random.h
@@ -1,5 +1,0 @@
-#pragma once
-#include "primitives.h"
-
-/// Returns the next pseudo-random number; the algo is very very simple
-i32 random_next();