+++ src/gcc/stdarg.h
@@ -133,4 +133,3 @@ typedef __gnuc_va_list va_list;
 
 #endif /* not _ANSI_STDARG_H_ */
 #endif /* not _STDARG_H */
-