+++ src/modern/integer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "primitives.h"
+#include "leper3.h"
 
 typedef struct {
     u8 x;