+++ src/random.c
@@ -1,4 +1,4 @@
-#include "random.h"
+#include "leper3.h"
 
 static i32 _random_last = 0;