+++ src/vmm.c
@@ -1,7 +1,6 @@
 // Virtual Memory Manager
 
-#include "page.h"
-#include "primitives.h"
+#include "leper3.h"
 
 typedef enum {
     vmm_Flag_present = 1,
@@ -12,19 +11,19 @@ void vmm_init() {
     // TODO use static memory (after allowing bigger kernel)
     u32 *directory = page_allocate();
 
-    for (address i = 0; i < 1024; i++) {
+    for (usize i = 0; i < 1024; i++) {
         directory[i] = vmm_Flag_rw;
     }
 
     u32 *kernel = page_allocate();
 
-    for (address i = 0; i < 1024; i++) {
+    for (usize i = 0; i < 1024; i++) {
         kernel[i] = (i * 4096) | vmm_Flag_present | vmm_Flag_rw;
     }
 
     kernel[0] = 0; // first page unmapped for null
 
-    directory[0] = ((address)kernel) | vmm_Flag_present | vmm_Flag_rw;
+    directory[0] = ((usize)kernel) | vmm_Flag_present | vmm_Flag_rw;
 
     __asm__ volatile("mov %0, %%cr3" : : "r"(directory));